Close the registry's fields; split it into a module
Two changes, both about what the registry shows.

**Fields are crate-internal.** `input_types` / `output_types` and the five
plan maps were `pub`. Outside the crate a table is now reached through
`Conversions::conversion` and `crossings` — which is what makes direction a
parameter rather than half of a field name, and what stops anyone observing
a cell before `supply` has graded it. `expansion_plans` gets an inherent
accessor (it is emission-only, so it stays off the `Conversions` trait, per
9fdbb9e); the other maps already had one. `pub(crate)` rather than private
because `expand` and `unfold` fill them — they are core's own state, just
not the world's.

**One 1989-line file becomes eleven, none over 420.** Grouped by what they
answer, not by type:

    mod         Registry, Declared, Decompositions, wiring
    key/cell    TypeKey; TypeSubject/TypeCell/TypeEntry/Direction
    declare     configure — every method records, none derives
    model       questions about the model
    scan        derive the crossing set
    order       hand the demand over, grade the answers
    run         prepare/finish/apply plans
    view        Conversions + Building
    error       what can go wrong
    walk        structural type-graph helpers

Inherent impls span modules, so `impl<M> Registry<M>` splits with them.

Two things the split surfaced rather than caused:

  * `TypeKeyParseError` and `DuplicateNameError` are reachable from the
    public API — `TypeKey::parse` returns one, `ScanError::DuplicateName`
    carries the other — and were never re-exported. Now they are, along
    with `NotExpressibleEntry`.
  * the boundary ledger moves 11 classification sites from `registry.rs` to
    `scan.rs` (2) + `walk.rs` (9). Total unchanged at 167: relocation, no
    new classifier.

594 tests, warning output byte-identical, regen-check clean, MSRV clippy
clean, doc warnings 44 -> 43.
