+++ prebindgen/src/api/core/mod.rs
@@ -42,7 +42,8 @@ pub use self::{
     niches::{NicheSlot, Niches},
     prebindgen::{const_path_alias, ConverterImpl, Prebindgen, Stage},
     registry::{
-        Building, Conversions, Crossing, Decompositions, Direction, Generation, Registry,
-        ScanError, TypeCell, TypeEntry, TypeKey, TypeSubject, WriteRustError,
+        Building, Conversions, Crossing, Decompositions, Direction, DuplicateNameError, Generation,
+        NotExpressibleEntry, Registry, ScanError, TypeCell, TypeEntry, TypeKey, TypeKeyParseError,
+        TypeSubject, WriteRustError,
     },
 };
