+++ prebindgen/src/api/core/registry/tests.rs
@@ -1,6 +1,7 @@
 use std::collections::HashSet;
 
 use proc_macro2::TokenStream;
+use quote::ToTokens;
 
 use super::*;
 use crate::api::core::{
