+++ prebindgen/src/lib.rs
@@ -300,9 +300,10 @@ pub mod core {
     /// where an adapter reaches for it.
     pub use crate::api::core::{
         warn_unclaimed, Building, Claimed, Conversions, ConverterImpl, Crossing, Decompositions,
-        Direction, DomainScalar, Element, Flat, Generation, Gravestone, NicheSlot, Niches,
-        Prebindgen, Registry, RepresentationDomain, ScalarValue, ScanError, Stage, Transmute,
-        TypeCell, TypeEntry, TypeKey, TypeSubject, WriteRustError,
+        Direction, DomainScalar, DuplicateNameError, Element, Flat, Generation, Gravestone,
+        NicheSlot, Niches, NotExpressibleEntry, Prebindgen, Registry, RepresentationDomain,
+        ScalarValue, ScanError, Stage, Transmute, TypeCell, TypeEntry, TypeKey, TypeKeyParseError,
+        TypeSubject, WriteRustError,
     };
 }
